Update dependencies in gradle
Updated the Kissen-Core-API and Kissen-Core-Source dependencies in the build.gradle.kts of both KissenPaper-Server and KissenPaper-API. Instead of using the versioned jar dependency, the project dependencies are now directly referenced. This change facilitates rapid development and testing as it eliminates the need to rebuild and publish the core libraries every time a change is made.