Integrated Velocity into the project & refactored methods in KissenPaper API
Added support for Velocity to the KissenPaper project. Updated several methods in the KissenPaper API, including refactoring the `getCommand` method to `obtainCommand`. The Context argument for some methods in the KissenPaperBanImplementation class, and some classes in the CraftOfflinePlayer and CraftPlayer classes are updated to be not nullable.

In the KissenPaperCommandImplementation class, adjusted setupComplete method to register the new TestCommand class and also renamed variables e.g. 'name' to 'value', etc. for better code clarity.

Furthermore, new file K