"Refactor build script and update clone task"
This commit includes improvements to our build.gradle.kts script. The 'mavenLocal' repository was removed to streamline our build process. Additionally, the process to clone from another repository was refactored for better readability and maintainability. This includes updating the 'cloneVelocity' task to use a more modern API and register it in a more idiomatic way for Gradle Kotlin DSL. The 'applyPatches' task has been updated to rely on 'cloneVelocity', ensuring the correct order of operations.