Implemented changes to enhance Ban functionality
In this commit, several changes have been made to enhance and extend the functionality of the Ban system. A new command `ban` is added in the class KissenPaperBanCommand, allowing users to ban or punish a specific user, thereby removing their access rights from the server.

The Ban Implementation has been updated to provide more debug information when retrieving a ban by its ID.

Two new classes have been created: `PaperBanParser` and `PaperBan`, for processing and handling Paper bans. These changes are essential for a reliable and functional banning system and will ensure users can be effectively managed and moderated.

Furthermore, getReason method in KissenPunishment class is renamed to getCause to be more appropriate and maintain consistency. Small modifications are made to handle NullPointerExceptions and enhance the object serialization.