"Refactor banning system and fix typos"
This update refactors the banning system by adding additional methods to the ban implementation that allows specifying whether to apply a ban immediately or not while invoking the punish method. Several methods related to getting the player's punishment details were renamed for clarity. For instance, the getPlayerBanSet methods have been changed to getPunishmentSet, to align better with the system's terminology. The serializer used in KissenPunishment.java was switched from a MiniSerializer to JsonSerializer, enhancing the serialization process's readability and compatibility. A typo in the multiplayer disconnect message has also been fixed. These changes collectively aim to increase code readability and maintainability.