Add UserRepository and update methods to align with new data structure
A new UserRepository interface is implemented to centralize actions related to user data retrieval. An additional method to get users by their name, not just UUID, is also introduced in this commit to enhance user-based operations. This requires changes to methods and paths in several classes which were adapted to interact with the new data model.

In addition, a semantic issue on the getBanHistory method is handled by renaming it to getPunishmentSet for better reflectance of its functionality. Minor change was applied to the equals method in KissenUser to compare SavableIDs instead of IDs, for better alignment with recent changes in project's data management.