Fix incorrect SQL query for retrieving stored data
Resolved an SQL syntax error in the KissenJDBCMeta file. A wrong column name was being referenced for the WHERE clause in the SELECT statement. The getKeyColumn() method is used now instead of the incorrect getValueColumn() method. Also introduced a DELETE statement before every INSERT to prevent data duplication when repeatedly setting a string for a specific key.

Signed-off-by: Ivo <chniucg@gmail.com>