"Implement custom chat renderer and settings in KissenChatImplementation"
In this commit, a custom renderer was implemented in KissenChatImplementation.java for message preparation. This function prepares messages from a server entity to another. A similar change was also made in ServerPlayer.java and CraftPlayer.java. This was necessary to ensure that the renderer is used whenever a message is sent using the system messaging functions.

In addition, user color settings were added in KissenUserImplementation.java. These settings allow a user to customize their primary, secondary, default, enabled, and disabled colors. This will give users more flexibility in personalizing their chat experience.

Finally, minor adjustments were made to the DefaultTheme.java, ChatImplementation.java, and Theme.java to align them with the new features.