Integrated ban system and localization
This commit integrates a ban system and localization into the KissenPaper server. The ban system was modified to include text color for each punishment's attributes and uses the sender's locale. Other changes are made to the server, including upgrading the text formatter to match the sender's format, and adding color to the ban cause.

The localization was also incorporated into the server. The updateLocale method was added to ServerPlayer, which sets the player's locale in the BukkitEntity. It is then used to open a book in the given locale.

Signed-off-by: ivo <chniucg@gmail.com>