Updated TranslatableComponent method to use fallback color
This commit modifies the 'transformTranslatableComponent' method in the 'DefaultTheme.java' file. Previously, it used a fixed default color for all translatable components. Now, it will attempt to use the translatable component's color if available, and if not, it will fall back to a provided fallback color. This change increases flexibility and avoids issues if the default color is not set.

Signed-off-by: ivo <chniucg@gmail.com>